+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 
 This document records the main changes to the ``sdssdb`` code.
 
+* Add method ``get_database_uri()`` to ``DatabaseConnection``.
+
 * :release:`0.12.0 <2024-06-28>`
 * Updated ``sdss_id_to_catalog`` view models to use double underscore separating table name from primary key.
 * Very significantly speed up the reflection of the database by caching the schema metadata.